# rest-client/openapi-definition.json
{
    "components": {
        "schemas": {
...
            "ErrorDto": {
                "properties": {
                    "error": {
                        "description": "Error string",
                        "type": "string"
                    },
                    "message": {
                        "description": "Message",
                        "type": "string"
                    },
                    "statusCode": {
                        "description": "HTTP status code",
                        "type": "number"
                    }
                },
                "required": [
                    "statusCode",
                    "message",
                    "error"
                ],
                "type": "object"
            },
            "ProductL0Dist": {
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "l0CoverageLocation": {
                        "type": "string"
                    },
                    "l0InstrumentFiles": {
                        "items": {
                            "$ref": "#/components/schemas/ProductL0InstrumentFile"
                        },
                        "type": "array"
                    },
                    "sourceProduct": {
                        "$ref": "#/components/schemas/ProductL0Src"
                    }
                },
                "required": [
                    "id",
                    "sourceProduct",
                    "l0InstrumentFiles",
                    "l0CoverageLocation"
                ],
                "type": "object"
            },
            "ProductL0DistDto": {
                "properties": {
                    "l0CoverageLocation": {
                        "description": "Location of coverage file for extent",
                        "type": "string"
                    }
                },
                "required": [
                    "l0CoverageLocation"
                ],
                "type": "object"
            },
            "ProductL0InstrumentFile": {
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "l0InstrumentFile": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "l0InstrumentFile"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "access-token": {
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "contact": {
            "email": "AusSeabed@ga.gov.au",
            "name": "AusSeabed",
            "url": "http://ausseabed.gov.au/"
        },
        "description": "The API description for the Ausseabed product catalogue inventory",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "title": "AusSeabed product catalogue",
        "version": "0.1.10"
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/products/l0-dist": {
            "get": {
                "operationId": "ProductsL0DistController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductL0Dist"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            },
            "post": {
                "operationId": "ProductsL0DistController_create",
                "parameters": [
                    {
                        "in": "query",
                        "name": "productL0SrcId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL0DistDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL0Dist"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            }
        },
        "/products/l0-dist/{productId}": {
            "delete": {
                "operationId": "ProductsL0DistController_delete",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            },
            "get": {
                "operationId": "ProductsL0DistController_findOne",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL0Dist"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            },
            "put": {
                "operationId": "ProductsL0DistController_update",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL0DistDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            }
        },
        "/products/l0-dist/{productId}/instrument-files": {
            "get": {
                "operationId": "ProductsL0DistController_findInstruments",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductL0InstrumentFile"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            },
            "post": {
                "operationId": "ProductsL0DistController_createInstrument",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL0InstrumentFileDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL0InstrumentFile"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            }
        },
        "/products/l0-dist/{productId}/instrument-files/{instrumentId}": {
            "delete": {
                "operationId": "ProductsL0DistController_deleteInstrument",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "path",
                        "name": "instrumentId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the instrument"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            },
            "get": {
                "operationId": "ProductsL0DistController_findOneInstrument",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "path",
                        "name": "instrumentId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/ProductL0InstrumentFile"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": ""
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            },
            "put": {
                "operationId": "ProductsL0DistController_updateInstrument",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "in": "path",
                        "name": "instrumentId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL0InstrumentFileDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the instrument"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l0-dist"
                ]
            }
        },
...
        "/products/l3-dist/{productId}": {
            "delete": {
                "operationId": "ProductsL3DistController_delete",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-dist"
                ]
            },
            "get": {
                "operationId": "ProductsL3DistController_findOne",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductL3Dist"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-dist"
                ]
            },
            "put": {
                "operationId": "ProductsL3DistController_update",
                "parameters": [
                    {
                        "in": "path",
                        "name": "productId",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductL3DistDto"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": "Could not find the survey"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Unable to authenticate request."
                    },
                    "408": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorDto"
                                }
                            }
                        },
                        "description": "Server took too long to respond."
                    }
                },
                "security": [
                    {
                        "access-token": []
                    }
                ],
                "tags": [
                    "products/l3-dist"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/rest"
        }
    ],
...
}